[update] veml6070
